+++ backend/posts/api/serializers.py
@@ -1,9 +1,14 @@
 from rest_framework import serializers
-from ..models import Client, FormData, FormDataCompany
+from ..models import Client, FormData, FormDataCompany, Company
 
 class ClientSerializer(serializers.ModelSerializer):
     class Meta :
         model = Client 
+        fields = ('name','variable1','variable2','variable3')
+
+class CompanySerializer(serializers.ModelSerializer):
+    class Meta :
+        model = Company 
         fields = ('name','variable1','variable2','variable3')
