+++ backend/posts/api/views.py
@@ -1,11 +1,15 @@
 from rest_framework.viewsets import ModelViewSet
-from ..models import Client
-from .serializers import ClientSerializer, FormDataSerializer, FormCompanyDataSerializer
+from ..models import Client, Company
+from .serializers import ClientSerializer, FormDataSerializer, FormCompanyDataSerializer, CompanySerializer
 from rest_framework.response import Response
 
 class ClientViewSet(ModelViewSet):
     queryset = Client.objects.all()
     serializer_class = ClientSerializer
+
+class CompanyViewSet(ModelViewSet):
+    queryset = Company.objects.all()
+    serializer_class = CompanySerializer
 
 def create(self, request, *args, **kwargs):
         serializer = FormDataSerializer(data=request.data)
